Remove miniscreen/core/fps_regulator.py
